Uav support

+++ src/main/java/jhi/germinate/server/resource/fileresource/FileResourceTypeResource.java
@@ -1,17 +1,18 @@
 package jhi.germinate.server.resource.fileresource;
 
+import jakarta.annotation.security.PermitAll;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.*;
 import jhi.germinate.resource.enums.UserType;
-import jhi.germinate.server.Database;
+import jhi.germinate.server.*;
 import jhi.germinate.server.database.codegen.tables.pojos.*;
 import jhi.germinate.server.database.codegen.tables.records.FileresourcetypesRecord;
 import jhi.germinate.server.resource.*;
 import jhi.germinate.server.util.*;
 import org.apache.commons.io.FileUtils;
-import org.jooq.DSLContext;
+import org.jooq.*;
 
-import jakarta.annotation.security.PermitAll;
-import jakarta.ws.rs.*;
-import jakarta.ws.rs.core.*;
+import java.io.File;
 import java.io.*;
 import java.sql.*;
 import java.util.List;
@@ -30,11 +31,17 @@ public class FileResourceTypeResource extends ContextResource
 	public List<ViewTableFileresourcetypes> getFileResourceType()
 		throws SQLException
 	{
+		AuthenticationFilter.UserDetails userDetails = (AuthenticationFilter.UserDetails) securityContext.getUserPrincipal();
+
 		try (Connection conn = Database.getConnection())
 		{
 			DSLContext context = Database.getContext(conn);
-			return context.selectFrom(VIEW_TABLE_FILERESOURCETYPES)
-						  .fetchInto(ViewTableFileresourcetypes.class);
+			SelectWhereStep<?> step = context.selectFrom(VIEW_TABLE_FILERESOURCETYPES);
+
+			if (!userDetails.isAtLeast(UserType.DATA_CURATOR))
+				step.where(VIEW_TABLE_FILERESOURCETYPES.PUBLIC_VISIBILITY.eq(true));
+
+			return step.fetchInto(ViewTableFileresourcetypes.class);
 		}
 	}
 
@@ -80,22 +87,35 @@ public class FileResourceTypeResource extends ContextResource
 		try (Connection conn = Database.getConnection())
 		{
 			DSLContext context = Database.getContext(conn);
-			// Delete all files associated with fileresource database objects.
-			File target = ResourceUtils.getFromExternal(resp, Integer.toString(fileResourceTypeId), "data", "download");
 
-			try
-			{
-				FileUtils.deleteDirectory(target);
-			}
-			catch (IllegalArgumentException e)
-			{
-				e.printStackTrace();
-			}
+			FileresourcetypesRecord record = context.selectFrom(FILERESOURCETYPES)
+													.where(FILERESOURCETYPES.ID.eq(fileResourceTypeId))
+													// Don't allow deletion of these new default resource types
+													.andNot(FILERESOURCETYPES.NAME.eq("Trials Shapefile").and(FILERESOURCETYPES.DESCRIPTION.eq("Shape file associated with a phenotypic trial. Fields within the shape file have to match the database entries.")))
+													.andNot(FILERESOURCETYPES.NAME.eq("Trials GeoTIFF").and(FILERESOURCETYPES.DESCRIPTION.eq("GeoTIFF file associated with a phenotypic trial. The \"created_on\" date of this fileresource determines the time point at which it was recorded.")))
+													.fetchAny();
 
-			// Delete the fileresource type. This will trigger the deletion of the referencing fileresouces.
-			return context.deleteFrom(FILERESOURCETYPES)
-						  .where(FILERESOURCETYPES.ID.eq(fileResourceTypeId))
-						  .execute() > 0;
+			if (record != null)
+			{
+				// Delete all files associated with fileresource database objects.
+				File target = ResourceUtils.getFromExternal(resp, Integer.toString(record.getId()), "data", "download");
+
+				try
+				{
+					FileUtils.deleteDirectory(target);
+				}
+				catch (IllegalArgumentException e)
+				{
+					e.printStackTrace();
+				}
+
+				// Delete the fileresource type. This will trigger the deletion of the referencing fileresouces.
+				return record.delete() > 0;
+			}
+			else
+			{
+				return false;
+			}
 		}
 		catch (IOException e)
 		{
